Plasmoid: Improve the list section delegate
This seems very much like a hack, but I cannot figure out a better way
to do this.

Add a forced margin of 80 pixels in the list results. This leaves a
considerable amount of space in the begining. The list section delegate
renders the text over there. Using anchors with the section delgate
doesn't seem to work. Only manual x,y coordinates seem to have any
effect, so I've used them.