Add a configuration widget and the concept of MatchTypes
Before this each MatchType was represented by one string. Now it has its
own class called MatchType. Each Match returns a MatchType* instead of
a string.

Each MatchType can be enabled or disabled based on the preferences. Each
source only returns the types which are enabled.

This patch implements all of this along with a configuration widget to
modify which types should be returned.