NepomukSource: Use a handmade query for everything else as well
It's somewhat faster than the one generated by the Nepomuk Query Library
as it doesn't have the FILTER(REGEX(?r, '^nepomuk')) in the end. The
FILTER is normally required in order to not get ontology results,
however since we are already requesting the results by type, it is not
required.