NepomukSource: Optimize email fetching
Fetch the subject along with the query, this way we avoid using the
Resource class and can skip the extra query per email shown.

This is especially useful because the Resource class loads all of the
email data. This includes the plain text content which is not at all
required.