PreviewPlugins: Let the plugin set the size
When the plugin creates a QDeclarativeItem, it can set the width/height
over there, and that will be respected. This is better than each preview
having the same dimensions.

Currently, both the email and text previews are kept at 384x384. But that's
just cause I haven't implemented the code to figure out the correct size
dynamically.