NepomukSource: When searching for Audio search for nmm:MusicPiece
This is not 100% accurate as not all nfo:Audio are nmm:MusicPieces. But
we cannot keep using nfo:Audio as almost all video are also of type
nfo:Audio. We could use a FILTER NOT EXISTS { ?r a nfo:Video . } but
that would slow things down even more. Also, virtuoso seems to give an
error with that kind of query.

So, for now we miss some of the search results, and Nepomuk should be
patched up to make all nfo:Audio -> nmm:MusicPiece.