Preview: Only load the previews after 500 msecs
This way moving the mouse over previews doesn't cause them to load and
block the UI. Ideally, the previews should never block but the Okular
one does.

This code required a big hack because Plasma::Dialog is strange and has
a hardcoded 150 msec timer in a certain place. It should be fixed.
Without our timer (of 160 msec) the dialog size is never updated and it
it is displayed as a small dot.