Add a timer for setting focus
The focus is not always set correctly. The hunch is that the focus
function is called before the popup is actually visible and
therfore the setFocus call does not do anything. So, we are using
a small timer to call the setTextFieldFocus function again.

Tested this over the course of 2 days, I haven't had an problems with
focus. If other people can confirm that this fixes the issue then we
look more into fixing this in plasma.