Always show the type on the top
We used to show the type of the result by hooking up to the ListView's
section delegate. This required a lot of hackery to get it to align
properly. It also resulted in the section being shown at the bottom when
the ListView was reversed. We always want it to be on the top.

We now show the type via the normal delegate. It is only shown when the
next index has a different type. This way we achieve the same result as
the section delegate but with less hacks.