SourcesModel: Re-order the types based on the "type" of the result
In this case the type corresponds to the type set by each runner which
can then be converted into an integer. The most commonly used types are
"Completion Match" (30) or "Exact Match" (100).

This is not perfect as applications are still not always on top, but it
is a step in the right direction, given that we're trying our best to be
all "generic".