SourcesModel: Clear the model immediately
We wait 250 msecs, so that the results appear in one go instead of each
signal being propogated as the results appear in a haphazzard fashion.
This works well for normal queries, but it also results in us waiting an
extra 250 msecs when the query is empty, and the data should be cleared.

Since the views typically already have a delay when passing the query
string, it makes sense to not wait.