The RunnerManager repeats the same results
It gives the results via a changed signal which contains ALL the
results, not just the new ones. This is why we were seeing a number of
duplicate entries in the list.

Right now, instead of checking each item against the entire list and
only updating the duplicates, we reset the entire model. It's simpler
this way (for now).