SourcesModel: Improve the sorting of the types
We used to just sort the results based on how we got them from the
krunner framework. The runner framework doesn't do much. It mostly just
gives the results as it gets them + some extra weight for results that
have been run before + some weight based on the type of the result
(Completion Result, Exact match, etc). This never resulted in
Applications always being on the top unless the full application name
was written.

This patch improves this by comparing the user visible text. The type
where the item's text contains the user visible string are given a
higher priority.