SourcesModel: Only consider each types first result for priority
Each type is given an additional priority if any of its shown results
has the string which the user searched. This is not perfect as often the
top most result of that type doesn't have that string but is still given
a higher priority. Arguably, the backends should also be improved.

BUG: 338228