Revert "SourcesModel: Only consider each types first result for priority"
This reverts commit 3e4bb6b1a7cd3df47b460cd0a7222605eeddd104.

Only considering the top most item is not ideal as the top most item
changes a lot based on which result you pick because krunner tries to
remember it.

Fortunately, the bug is not re-introduced as we did add the extra
sorting based on if the visible text contains the query.