Support both QString and QIcon for dragHelper pixmap
The model was changed a while ago to return a QString if possible to have icon processing done in the delegate QML on-demand.
This broke the drag helper's drag pixmap, however, as the QML would pass in model.decoration which might be a string.

CHANGELOG: Fixed drag pixmap not showing when dragging results off KRunner's window

Differential Revision: https://phabricator.kde.org/D14545