Fix current index behavior with mouse and keyboard
When the mouse wasn't used (!moved) at all, ignore any mouse events to avoid the currentIndex changing
just because the mouse happened to be underneath the results (current behavior).
Once the mouse was used, upate the index live.
Now that we have a proper model, we get proper insertions, so ListView will take care of updating
the currentIndex accordingly and shift it around.
We don't want that happening for the topmost item, though, which is the one selected if the user hasn't
done any keyboard navigation.

With this change, results coming in delayed won't unexpectedly reset the currentIndex