Fix running results automatically
runCurrentIndex() relies on the ListView being up to date.
However, all we end up doing is running the first result, so we might as well
bypass ListView and call the model directly.
Also, quit the match session before beginResetModel(), otherwise we might get
results added while the model is in the process of resetting, causing crashes