Size category width dynamically
In case all the category names are short, this leaves more
room for the result text.

To reduce jumpiness of the UI, the category width is only
reset when clearing the query. As a result, sometimes you
still get a wide category column when you e.g. had a
transient "Recently used file" result earlier in your query