Fix mouse interaction with history entries
In case of the history entries we do not have
a ResultView, but rather a normal QML list view.
Consequently special properties/methods do not
exist.

This has the only downside, that one can accidentally
focus one item with the mouse when the user tries to do
a keyboard only interaction. But considering that clearing
specific entries from the history is done rarely and mostly
done by clearing all entries in the KCM this should be acceptable.

BUG: 431010
FIXED-IN: 5.21