Make results delegate list items a bit taller with bigger icons
We generally use the smallMedium (22px) icon size for
small-ish-but-not-too-small icons as they have a better visual balance
with adjacent text, compared to the small (16px) icon size. 22px icons
also generally look better than 16px icons. Moving the results list
items to use smallMedium icons offers a better visual balance, and
adding a tiny of vertical padding  balances out everything a bit more
to provide more "breathing room" between the rows.

As a consequence, the maximum number of results visible at once has to
decrease a bit, but this is generally not a major problem as the user
can simply keep typing to refine their search to find what it is they're
looking for.

BUG: 422567
FIXED-IN: 5.23