Remove check if result conains query string
This essentially overrides the runner relevance and match type.

An example is my custom runner for the nordvpn cli, having the query string in the
result text would be unneeded, because one already has the category set.

This is also a perquisite for https://bugs.kde.org/show_bug.cgi?id=340283

Before:
https://phabricator.kde.org/file/data/jjdng2qroopr5w6sv2mq/PHID-FILE-yvv5nhamspryvbbkfwd4/Screenshot_20220508_130532.png

After:
https://phabricator.kde.org/file/data/lop7apvdb2tn276ngk5k/PHID-FILE-2yxwbxbayxkn54rtlzth/Screenshot_20220508_130702.png