lib: simplify mouse hover handling
ResultsView/ResultDelegate do some complicated mouse interaction
handling: the delegate listens to a HoverHandler's pointChanged signal.
which updates in a lot of circumstances, and then filters based on
keeping track whether the mouse moved to not activate with keyboard-only
use. This doesn't work well if there is mixed keyboard/mouse use, where
arrow key navigation becomes useless almost as soon as the mouse is
involved. And it isn't really necessary in the first place, as the real
problem is connecting to such a busy signal.

Instead, we drop the whole state-tracking jazz and even the
HoverHandler, and just listen to the delegate's own hovered property.
hovered only updates after the mouse moves, so we get essentially
the same thing for free but without the edge case bugs.

BUG: 501350
FIXED-IN: 6.4.0