plasmoid: set layout minimum size on full representation
The Plasmoid's full representation computes the size it should use, but
does not actually apply it. If the user adds it to a panel, it will
default to the search field's size, and not update as the user types a
search so the results are invisible. While the user can manually
increase the size (if they happen to figure it out), that means the
applet will no longer shrink when the search is cleared. The minimum and
maximum sizes also apply to the compactRepresentation, which causes
panel applets to resize in weird ways in some circumstances.

Instead, set the computed sizes as the default values, and get a similar
effect to the regular krunner popup but as a panel applet. Also remove
the size changes from the PlasmoidItem and only apply them to the
fullRepresentation.