plasmoid: simplify Layout change
The applet has a reversed orientation on bottom panels, with the search
field below the results. This allows a more consistent position when the
content changes. To do this, it reorders the children property of the
layout and futzes around with signals on edge changes, which feels a
little messy.

Instead, we can just use LayoutItemProxies to move the search field to
the correct position and have property binding take care of everything
else automatically. This is simpler and more declaratice. Along the way,
we turn the check for bottom edge from a repeatedly called function into
a simple property.