resultsview: refactor move_category so it doesn't deadlock
- shared impl for both move directions
- have a finite loop condition
- break the loop when finding the first entry of a category

(slightly complicated because the break condition is divergent, I've
opted to handle this inside the move function. another alternative would
have been to pass in an extra conditions function, but that seems just
as complicated and makes for less readable code)

the secret to this function working for both scenarios actually lies in
the caller. the caller suplies a list of indexes and those are either
ascending or descending sorted. effectively allowing us to iterate
always in one direction in the move function.

BUG: 509786


(cherry picked from commit cac8ff46a4f67480ea532399ea5dd26aac7d9d8f)

Co-authored-by: Harald Sitter <sitter@kde.org>