Port ResultDelegate to required properties
A bit unfun since we expect the model to have a 'display' role
but that's already a property in ItemDelegate. To work around this
use "required property var model"

Also clean up some 'interesting' role handling.

additionalActions is not needed any more since actions is a proper
property now.

Introduce a displayText property so that the text can be overridden
by the consumer (used in the KRunner history feature)